+++ src/collective/elastic/ingest/__init__.py
@@ -1,4 +1,0 @@
-import os
-
-
-OPENSEARCH = os.environ.get("INDEX_OPENSEARCH") == "1"