+++ CHANGES.rst
@@ -6,7 +6,10 @@ Changelog
 
 - Fix: Plone schema caching logic was broken. [jensens]
 - Fix: Remove unnecessary ``batching`` on preprocessing. [jensens]
-- Fix: Add missigit sng ``zope.schema._field.Set`` to examples mapping. [jensens]
+- Fix: Add missing ``zope.schema._field.Set`` to examples mapping. [jensens]
+- Enhancement: Move vocabulary and section handling to preprocessing.
+  Turn ``ingest`` package into module. [jensens]
+- Fix: Get OPENSEARCH from c.e.ingest module and do not dup here. [jensens]
 
 2.0.0rc4 (2023-11-28)
 ---------------------